#49 implement override csv option

+++ R/drug_atc_codes_node_parser.R
@@ -69,6 +69,11 @@ get_atc_codes_df <- function(rec) {
 #' }
 #' @export
 parse_drug_atc_codes <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_atc_codes", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_atc_codes <- map_df(pkg.env$children, ~ get_atc_codes_df(.x)) %>% unique()
 
   write_csv(drug_atc_codes, save_csv, csv_path)
