+++ R/drug_carrier_node_parser.R
@@ -104,6 +104,11 @@ get_carriers_polypeptide_go_classifiers_df <- function(rec) {
 #' }
 #' @export
 parse_drug_carriers_actions <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_carriers_actions", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_carriers_actions <-
     map_df(pkg.env$children, ~ get_carriers_actions_df(.x)) %>% unique()
 
@@ -175,6 +180,11 @@ parse_drug_carriers_actions <- function(save_table = FALSE, save_csv = FALSE, cs
 #' }
 #' @export
 parse_drug_carriers_articles <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_carriers_articles", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_carriers_articles <-
     map_df(pkg.env$children, ~ get_carriers_articles_df(.x)) %>% unique()
 
@@ -242,6 +252,11 @@ parse_drug_carriers_articles <- function(save_table = FALSE, save_csv = FALSE, c
 #' }
 #' @export
 parse_drug_carriers_textbooks <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_carriers_textbooks", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_carriers_textbooks <-
     map_df(pkg.env$children, ~ get_carriers_textbooks_df(.x)) %>% unique()
 
@@ -307,6 +322,11 @@ parse_drug_carriers_textbooks <- function(save_table = FALSE, save_csv = FALSE, 
 #' }
 #' @export
 parse_drug_carriers_links <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_carriers_links", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_carriers_links <-
     map_df(pkg.env$children, ~ get_carriers_links_df(.x)) %>% unique()
 
@@ -375,6 +395,11 @@ parse_drug_carriers_links <- function(save_table = FALSE, save_csv = FALSE, csv_
 #' }
 #' @export
 parse_drug_carriers_polypeptides <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_carriers_polypeptides", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_carriers_polypeptides <-
     map_df(pkg.env$children, ~ get_carriers_polypeptide_df(.x)) %>% unique()
 
@@ -464,6 +489,11 @@ parse_drug_carriers_polypeptides <- function(save_table = FALSE, save_csv = FALS
 #' @export
 parse_drug_carriers_polypeptides_external_identifiers <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+    path <- get_dataset_full_path("drug_carriers_polypeptide_external_identifiers", csv_path)
+    if (override_csv & file.exists(path)) {
+      return(readr::read_csv(path))
+    }
+
     drug_carriers_polypeptide_external_identifiers <-
       map_df(pkg.env$children,
              ~ get_carriers_polypeptide_external_identifiers_df(.x)) %>% unique()
@@ -533,6 +563,11 @@ parse_drug_carriers_polypeptides_external_identifiers <-
 #' @export
 parse_drug_carriers_polypeptides_synonyms <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+    path <- get_dataset_full_path("drug_carriers_polypeptide_synonyms", csv_path)
+    if (override_csv & file.exists(path)) {
+      return(readr::read_csv(path))
+    }
+
     drug_carriers_polypeptide_synonyms <-
       map_df(pkg.env$children,
              ~ get_carriers_polypeptide_synonyms_df(.x)) %>%
@@ -603,6 +638,12 @@ parse_drug_carriers_polypeptides_synonyms <-
 #' @export
 parse_drug_carriers_polypeptides_pfams <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+    path <-
+      get_dataset_full_path("drug_carriers_polypeptide_pfams", csv_path)
+    if (override_csv & file.exists(path)) {
+      return(readr::read_csv(path))
+    }
+
     drug_carriers_polypeptide_pfams <-
       map_df(pkg.env$children,
              ~ get_carriers_polypeptide_pfams_df(.x)) %>%
@@ -674,6 +715,12 @@ parse_drug_carriers_polypeptides_pfams <-
 #' @export
 parse_drug_carriers_polypeptides_go_classifiers <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+    path <-
+      get_dataset_full_path("drug_carriers_polypeptides_go_classifiers", csv_path)
+    if (override_csv & file.exists(path)) {
+      return(readr::read_csv(path))
+    }
+
     drug_carriers_polypeptides_go_classifiers <-
       map_df(pkg.env$children,
              ~ get_carriers_polypeptide_go_classifiers_df(.x)) %>% unique()
@@ -740,6 +787,12 @@ parse_drug_carriers_polypeptides_go_classifiers <-
 #' }
 #' @export
 parse_drug_carriers <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <-
+    get_dataset_full_path("drug_carriers", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_carriers <- map_df(pkg.env$children, ~ get_carriers_df(.x)) %>% unique()
 
   write_csv(drug_carriers, save_csv, csv_path)
