+++ R/drug_parsers.R
@@ -47,6 +47,11 @@
 #' }
 #' @export
 parse_drug <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drugs", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   # db connection
   drugs <- map_df(pkg.env$children, ~ drug_df(.x)) %>% unique()
 
@@ -134,6 +139,11 @@ parse_drug <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", ove
 #' }
 #' @export
 parse_drug_groups <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_groups", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_groups <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "groups")) %>% unique()
 
@@ -200,6 +210,11 @@ parse_drug_groups <- function(save_table = FALSE, save_csv = FALSE, csv_path = "
 #' }
 #' @export
 parse_drug_articles <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_articles", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_articles <-
     map_df(
       pkg.env$children,
@@ -266,6 +281,11 @@ parse_drug_articles <- function(save_table = FALSE, save_csv = FALSE, csv_path =
 #' }
 #' @export
 parse_drug_books <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_books", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_books <-
     map_df(
       pkg.env$children,
@@ -331,6 +351,11 @@ parse_drug_books <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".
 #' }
 #' @export
 parse_drug_links <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_links", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_links <-
     map_df(
       pkg.env$children,
@@ -397,6 +422,11 @@ parse_drug_links <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".
 #' }
 #' @export
 parse_drug_synonyms <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_synonyms", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_synonyms <- map_df(pkg.env$children, ~ get_synonyms_df(.x)) %>% unique()
 
   write_csv(drug_synonyms, save_csv, csv_path)
@@ -461,6 +491,11 @@ parse_drug_synonyms <- function(save_table = FALSE, save_csv = FALSE, csv_path =
 #' }
 #' @export
 parse_drug_products <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_products", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_products <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "products")) %>% unique()
 
@@ -526,6 +561,12 @@ parse_drug_products <- function(save_table = FALSE, save_csv = FALSE, csv_path =
 #' @export
 parse_drug_calculated_properties <- function(save_table = FALSE,
                                              save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <-
+    get_dataset_full_path("drug_calculated_properties", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_calculated_properties <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "calculated-properties")) %>% unique()
 
@@ -588,6 +629,12 @@ parse_drug_calculated_properties <- function(save_table = FALSE,
 #' }
 #' @export
 parse_drug_international_brands <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <-
+    get_dataset_full_path("drug_international_brands", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_international_brands <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "international-brands")) %>%
     unique()
@@ -651,6 +698,12 @@ parse_drug_international_brands <- function(save_table = FALSE, save_csv = FALSE
 #' }
 #' @export
 parse_drug_salts <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <-
+    get_dataset_full_path("drug_salts", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_salts <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "salts")) %>%
     unique()
@@ -714,6 +767,11 @@ parse_drug_salts <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".
 #' }
 #' @export
 parse_drug_mixtures <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_mixtures", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_mixtures <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "mixtures")) %>% unique()
 
@@ -776,6 +834,11 @@ parse_drug_mixtures <- function(save_table = FALSE, save_csv = FALSE, csv_path =
 #' }
 #' @export
 parse_drug_packagers <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_packagers", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_packagers <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "packagers")) %>% unique()
 
@@ -839,6 +902,11 @@ parse_drug_packagers <- function(save_table = FALSE, save_csv = FALSE, csv_path 
 #' }
 #' @export
 parse_drug_categories <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_categories", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_categories <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "categories")) %>% unique()
 
@@ -901,6 +969,11 @@ parse_drug_categories <- function(save_table = FALSE, save_csv = FALSE, csv_path
 #' }
 #' @export
 parse_drug_affected_organisms <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_affected_organisms", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_affected_organisms <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "affected-organisms")) %>% unique()
 
@@ -967,6 +1040,11 @@ parse_drug_affected_organisms <- function(save_table = FALSE, save_csv = FALSE, 
 #' }
 #' @export
 parse_drug_dosages <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_dosages", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_dosages <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "dosages")) %>% unique()
 
@@ -1030,6 +1108,11 @@ parse_drug_dosages <- function(save_table = FALSE, save_csv = FALSE, csv_path = 
 #' }
 #' @export
 parse_drug_ahfs_codes <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_ahfs_codes", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_ahfs_codes <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "ahfs-codes")) %>% unique()
   if (nrow(drug_ahfs_codes) > 0) {
@@ -1095,6 +1178,11 @@ parse_drug_ahfs_codes <- function(save_table = FALSE, save_csv = FALSE, csv_path
 #' }
 #' @export
 parse_drug_pdb_entries <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_pdb_entries", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_pdb_entries <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "pdb-entries")) %>% unique()
 
@@ -1161,6 +1249,11 @@ parse_drug_pdb_entries <- function(save_table = FALSE, save_csv = FALSE, csv_pat
 #' }
 #' @export
 parse_drug_patents <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_patents", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_patents <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "patents")) %>% unique()
 
@@ -1224,6 +1317,11 @@ parse_drug_patents <- function(save_table = FALSE, save_csv = FALSE, csv_path = 
 #' }
 #' @export
 parse_drug_food_interactions <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_food_interactions", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_food_interactions <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "food-interactions")) %>% unique()
 
@@ -1291,6 +1389,12 @@ parse_drug_food_interactions <- function(save_table = FALSE, save_csv = FALSE, c
 #' }
 #' @export
 parse_drug_interactions <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_drug_interactions", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
+
   drug_drug_interactions <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "drug-interactions")) %>% unique()
 
@@ -1354,6 +1458,11 @@ parse_drug_interactions <- function(save_table = FALSE, save_csv = FALSE, csv_pa
 #' }
 #' @export
 parse_drug_experimental_properties <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_experimental_properties", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_experimental_properties <-
     map_df(pkg.env$children,
            ~ drug_sub_df(.x, "experimental-properties")) %>% unique()
@@ -1418,6 +1527,11 @@ parse_drug_experimental_properties <- function(save_table = FALSE, save_csv = FA
 #' }
 #' @export
 parse_drug_external_identifiers <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_external_identifiers", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_external_identifiers <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "external-identifiers"))
 
@@ -1480,6 +1594,11 @@ parse_drug_external_identifiers <- function(save_table = FALSE, save_csv = FALSE
 #' }
 #' @export
 parse_drug_external_links <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_external_links", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_external_links <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "external-links")) %>% unique()
 
@@ -1542,6 +1661,11 @@ parse_drug_external_links <- function(save_table = FALSE, save_csv = FALSE, csv_
 #' }
 #' @export
 parse_drug_snp_effects <- function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+  path <- get_dataset_full_path("drug_snp_effects", csv_path)
+  if (override_csv & file.exists(path)) {
+    return(readr::read_csv(path))
+  }
+
   drug_snp_effects <-
     map_df(pkg.env$children, ~ drug_sub_df(.x, "snp-effects")) %>% unique()
 
@@ -1606,6 +1730,11 @@ parse_drug_snp_effects <- function(save_table = FALSE, save_csv = FALSE, csv_pat
 #' @export
 parse_drug_snp_adverse_drug_reactions <-
   function(save_table = FALSE, save_csv = FALSE, csv_path = ".", override_csv = FALSE) {
+    path <- get_dataset_full_path("drug_snp_adverse_drug_reactions", csv_path)
+    if (override_csv & file.exists(path)) {
+      return(readr::read_csv(path))
+    }
+
     drug_snp_adverse_drug_reactions <-
       map_df(pkg.env$children,
              ~ drug_sub_df(.x, "snp-adverse-drug-reactions")) %>% unique()
